requirements.txt
torch
torchvision
numpy
trimesh
Pillow
matplotlib
ezdxf
pyrender
pytest
